requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
arrow==1.2.3
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
blessed==1.20.0
boto3==1.26.105
botocore==1.29.105
bravado==11.0.3
bravado-core==5.17.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmaes==0.9.1
cmake==3.26.1
colorlog==6.7.0
contourpy==1.0.7
croniter==1.3.14
cycler==0.11.0
dateutils==0.6.12
deepdiff==6.3.0
deepspeed==0.8.3
docstring-parser==0.15
fastapi==0.88.0
filelock==3.10.7
flatten-dict==0.4.2
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
GPUtil==1.4.0
greenlet==2.0.2
h11==0.14.0
hjson==3.1.0
hydra-core==1.3.2
idna==3.4
importlib-resources==5.12.0
inquirer==3.1.3
isoduration==20.11.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonargparse==4.20.1
jsonpointer==2.3
jsonref==1.1.0
jsonschema==4.17.3
kiwisolver==1.4.4
lightning==2.0.1.post0
lightning-cloud==0.5.33
lightning-utilities==0.8.0
lit==16.0.0
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
neptune-client==1.1.1
networkx==3.0
ninja==1.11.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
optuna==3.1.0
ordered-set==4.1.0
packaging==23.0
pandas==2.0.0
Pillow==9.4.0
protobuf==3.20.3
psutil==5.9.4
py-cpuinfo==9.0.0
pydantic==1.10.7
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.1
pytz==2023.3
PyYAML==6.0
readchar==4.0.5
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==13.3.3
s3transfer==0.6.0
simplejson==3.18.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.8
starlette==0.22.0
starsessions==1.3.0
swagger-spec-validator==3.0.3
sympy==1.11.1
tensorboardX==2.6
torch==2.0.0
torchmetrics==0.11.4
torchvision==0.15.1
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typeshed-client==2.2.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.15
uvicorn==0.21.1
wcwidth==0.2.6
webcolors==1.13
websocket-client==1.5.1
websockets==11.0.1
yarl==1.8.2
